+++ tic-tac-toe-app/src/components/Game.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { Board } from '../components/Board'
+
+export class Game extends React.Component {
+  render() {
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board />
+        </div>
+        <div className="game-info">
+          <div>{/* status */}</div>
+          <ol>{/* TODO */}</ol>
+        </div>
+      </div>
+    );
+  }
+};
