+++ tic-tac-toe-app/src/types/SquareValue.ts
@@ -1,0 +1,1 @@
+export type SquareValue = 'O' | 'X' | null;
